backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1